Clean up the dye color processors and add support for beds.
